+++ src/app/(root)/(routes)/mypage/mycards/components/my-card-list-content/index.tsx
@@ -1,0 +1,3 @@
+import MyCardListContent from './MyCardListContent'
+
+export default MyCardListContent
